+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/angular
 
+## 2.0.67
+
+### Patch Changes
+
+- Updated dependencies [53f6731]
+  - ai@6.0.67
+  - @ai-sdk/provider-utils@4.0.13
+
+## 2.0.66
+
+### Patch Changes
+
+- Updated dependencies [96936e5]
+  - @ai-sdk/provider-utils@4.0.12
+  - ai@6.0.66
+
+## 2.0.65
+
+### Patch Changes
+
+- ai@6.0.65
+
 ## 2.0.64
 
 ### Patch Changes