+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/langchain
 
+## 2.0.72
+
+### Patch Changes
+
+- Updated dependencies [53f6731]
+  - ai@6.0.67
+
+## 2.0.71
+
+### Patch Changes
+
+- ai@6.0.66
+
+## 2.0.70
+
+### Patch Changes
+
+- ai@6.0.65
+
 ## 2.0.69
 
 ### Patch Changes