+++ packages/react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/react
 
+## 3.0.69
+
+### Patch Changes
+
+- Updated dependencies [53f6731]
+  - ai@6.0.67
+  - @ai-sdk/provider-utils@4.0.13
+
+## 3.0.68
+
+### Patch Changes
+
+- Updated dependencies [96936e5]
+  - @ai-sdk/provider-utils@4.0.12
+  - ai@6.0.66
+
+## 3.0.67
+
+### Patch Changes
+
+- ai@6.0.65
+
 ## 3.0.66
 
 ### Patch Changes